+++ ichnaea/models/cell.py
@@ -270,7 +270,6 @@ class Cell(CellMixin, StationMixin, CreationMixin, _Model):
         Index('cell_created_idx', 'created'),
         Index('cell_modified_idx', 'modified'),
         Index('cell_new_measures_idx', 'new_measures'),
-        Index('cell_total_measures_idx', 'total_measures'),
     )
     _valid_schema = ValidCellSchema